+++ src/main/java/com/beforehairshop/demo/hairdesigner/repository/HairDesignerRepository.java
@@ -5,9 +5,10 @@ import com.beforehairshop.demo.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface HairDesignerRepository extends JpaRepository<HairDesigner, Long> {
+public interface HairDesignerRepository extends JpaRepository<HairDesigner, BigInteger> {
     List<HairDesigner> findAllByMember(Member member);
 }
