+++ src/main/java/com/beforehairshop/demo/hairdesigner/repository/HairDesignerWorkingDayRepository.java
@@ -5,9 +5,10 @@ import com.beforehairshop.demo.hairdesigner.domain.HairDesignerWorkingDay;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface HairDesignerWorkingDayRepository extends JpaRepository<HairDesignerWorkingDay, Long> {
+public interface HairDesignerWorkingDayRepository extends JpaRepository<HairDesignerWorkingDay, BigInteger> {
     List<HairDesignerWorkingDay> findAllByHairDesigner(HairDesigner hairDesigner);
 }
